Fix SwiftLint type/function body length violations blocking the build
Split YouTubePlayerView into +Background (scene-phase / audio-session
helpers) and +Helpers (computed properties, feedAvatarView) extensions
to bring the struct body under the 350-line limit.

Suppress the pre-existing type_body_length violation on
ScrollExpandedArticleView (unchanged by this PR) and the
function_body_length violation on ArticleExtractor.collectBlocks
(already had cyclomatic_complexity suppressed) so the build passes.
