+++ SakuraRSS/Views/Shared/YouTube Player/YouTubePlayerView.swift
@@ -1,4 +1,3 @@
-import AVFoundation
 import SwiftUI
 import WebKit
 import FoundationModels
@@ -22,12 +21,12 @@ struct YouTubePlayerView: View {
     @State private var hasStartedPlaying = false
     @State private var isPiP = false
     @State private var videoAspectRatio: CGFloat = 16 / 9
-    @State private var feed: Feed?
-    @State private var favicon: UIImage?
-    @State private var acronymIcon: UIImage?
+    @State var feed: Feed?
+    @State var favicon: UIImage?
+    @State var acronymIcon: UIImage?
     @State var chapters: [YouTubeChapter] = []
-    @State private var wantsPlaybackInBackground = false
-    @State private var playerID = UUID()
+    @State var wantsPlaybackInBackground = false
+    @State var playerID = UUID()
 
     // SponsorBlock
     @AppStorage("YouTube.SponsorBlock.Enabled") var sponsorBlockEnabled = false
@@ -48,98 +47,6 @@ struct YouTubePlayerView: View {
     @State var hasCachedSummary = false
     @State var showingSummary = false
     @State var summarizationError: String?
-
-    var isAppleIntelligenceAvailable: Bool {
-        SystemLanguageModel.default.availability == .available
-    }
-
-    var descriptionSource: String? {
-        article.summary ?? article.content
-    }
-
-    var hasDescription: Bool {
-        guard let descriptionSource else { return false }
-        return !descriptionSource.isEmpty
-    }
-
-    var hasTranslationForCurrentMode: Bool {
-        if showingSummary {
-            return translatedSummary != nil
-        }
-        return translatedText != nil || hasCachedTranslation
-    }
-
-    private var displayDescription: String? {
-        if showingSummary, let summarizedText {
-            if showingTranslation, let translatedSummary {
-                return translatedSummary
-            }
-            return summarizedText
-        }
-        if showingTranslation, let translatedText {
-            return translatedText
-        }
-        return descriptionSource
-    }
-
-    private func activateBackgroundAudioSession() {
-        let session = AVAudioSession.sharedInstance()
-        try? session.setCategory(.playback, mode: .moviePlayback)
-        try? session.setActive(true)
-    }
-
-    private func handleScenePhaseChange(_ newPhase: ScenePhase) {
-        switch newPhase {
-        case .background, .inactive:
-            wantsPlaybackInBackground = isPlaying
-            activateBackgroundAudioSession()
-            if isPlaying {
-                resumePlaybackIfNeeded()
-            }
-        case .active:
-            activateBackgroundAudioSession()
-            if wantsPlaybackInBackground {
-                resumePlaybackIfNeeded()
-                wantsPlaybackInBackground = false
-            }
-        @unknown default:
-            break
-        }
-    }
-
-    private func resumePlaybackIfNeeded() {
-        let script = """
-        (function() {
-            var v = document.querySelector('video');
-            if (v && v.paused) { v.play().catch(function(){}); }
-        })();
-        """
-        webView?.evaluateJavaScript(script, completionHandler: nil)
-    }
-
-    var youtubeAppURL: URL? {
-        guard let url = URL(string: article.url),
-              var components = URLComponents(url: url, resolvingAgainstBaseURL: false) else {
-            return nil
-        }
-        components.scheme = "youtube"
-        return components.url
-    }
-
-    @ViewBuilder
-    private var feedAvatarView: some View {
-        if let favicon {
-            FaviconImage(favicon, size: 36, circle: true, skipInset: true)
-        } else if let acronymIcon {
-            FaviconImage(acronymIcon, size: 36, circle: true, skipInset: true)
-        } else if let feed {
-            InitialsAvatarView(feed.title, size: 36, circle: true)
-        } else {
-            Circle()
-                .fill(.secondary.opacity(0.2))
-                .frame(width: 36, height: 36)
-        }
-    }
 
     var body: some View {
         VStack(spacing: 0) {
